requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-hashers-passlib==0.4
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.8
oauthlib==3.1.0
passlib==1.7.2
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
RestAuthCommon==0.7.0
six==1.14.0
South==1.0.2
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
